fix(speech): restore PTT with focused terminals; split oversized modules
- the widget-focus guard wrongly disabled push-to-talk in the normal
  case (terminal bodies hold egui focus): replaced with a text-surface
  gate (settings/palette/search/rename open) applied to presses only,
  so releases still stop an in-flight hold
- speech config types + validation moved to horizon-core/src/
  speech_config.rs (config re-exports keep all paths stable); speech
  settings UI moved to settings/speech.rs; lifecycle's event injection
  extracted — general.rs, config.rs, and the handler are back inside
  the module-size guardrails
