+++ crates/horizon-core/src/lib.rs
@@ -21,6 +21,7 @@ mod runtime_state;
 pub mod search;
 mod session_store;
 mod shortcuts;
+pub mod speech_config;
 pub mod speech_model;
 mod ssh;
 mod terminal;
